fix: ExitImpersonation access denied + remove mobile dashboard auto-redirect (CHANGE-013)
ExitImpersonation moved from AuthorController to AccountController so it
is reachable while signed in as BetaReader during impersonation. The
class-level RequireAuthorPolicy was blocking the action before it ran.

Removed the Kindle-style auto-redirect from MobileDashboard that was
sending users directly to their last read scene. The Continue Reading
CTA on the dashboard is the correct replacement for both desktop and
mobile.
